universityapi/lecturers/serializers.py
from rest_framework import serializers
from .models import Lecturer

class LecturerSerializer(serializers.ModelSerializer):
    class Meta:
        model = Lecturer
        fields = ('id', 'name', 'description')
